requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0
bcrypt==4.0.1 